refactor: replace @NoArgsConstructor and @Getter with @Value in DTO classes test: add unit tests for ParameterService to validate parameter operations
