+++ config-service/src/main/java/com/ayd/config_service/parameters/dtos/ParameterNITRequestDTO.java
@@ -4,9 +4,9 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Value;
 
-@NoArgsConstructor
-@Getter
+@Value
 public class ParameterNITRequestDTO {
     @NotBlank(message = "El nuevo tipo de regimen es obligatorio")
     @Pattern(regexp = "\\d+", message = "El NIT debe ser un valor numerico")
