+++ config-service/src/main/java/com/ayd/config_service/parameters/dtos/ParameterNombreRequestDTO.java
@@ -3,9 +3,9 @@ package com.ayd.config_service.parameters.dtos;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Value;
 
-@NoArgsConstructor
-@Getter
+@Value
 public class ParameterNombreRequestDTO {
     @NotBlank(message = "El nuevo nombre de la empresa es obligatorio.")
     private String newName;
